feat: add database migration scripts to restructure order items and update plant category references
