+++ backend/scripts/fix_categories.js
@@ -5,7 +5,7 @@ dns.setDefaultResultOrder('ipv4first');
 if (process.env.NODE_ENV !== "production") {
     dns.setServers(["8.8.8.8", "8.8.4.4"]);
 }
-const Plant = require('./src/model/nurseryModel/plants');
+const Plant = require('../src/model/nurseryModel/plants');
 
 const fixStringObjectIds = async () => {
     try {
